Remove wmmgr.h from wmframe.h. Add classHint() to ClientData.
