+++ src/wmtaskbar.cc
@@ -9,6 +9,7 @@
 #include "config.h"
 #include "wmtaskbar.h"
 #include "wmframe.h"
+#include "wmmgr.h"
 #include "wmconfig.h"
 #include "wmprog.h"
 #include "wmwinmenu.h"
